% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoc_fused.R
\name{fusedepoc}
\alias{fusedepoc}
\title{Fused EPoC}
\usage{
fusedepoc(Y, U, lambda1, lambda2, method = "G", trace = 1,
  optim = TRUE, stabilise = FALSE)
}
\arguments{
\item{Y}{list of K mRNA expression level matrices
classes of size N x p}

\item{U}{list of K Copy Number Aberration (CNA) matrices
of size N x p}

\item{lambda1}{LASSO penalty parameter lambda1 penalizes
the L1 norm of the resulting coefficients. This regulates
sparsity}

\item{lambda2}{Fused LASSO penalty parameter lambda2
penalizes the L1 norm of the difference between
coefficients in the K different classes}

\item{method}{'G' or 'A' as defined for the EPoC method.
In short, 'G' uses CNA as predictors while 'A' uses mRNA as
predictors.}

\item{trace}{The verbosity of information while running. 0
is quiet. Default 1.}

\item{optim}{Whether to use the optimized version or not}

\item{stabilise}{Whether to use the 10-times stabilization}
}
\description{
From several (K) data sets with data-levels mRNA and CNA
estimate a multi-graph representing gene regulatory networks with fusing showing similarities and differences between sets of cancer tumour samples
}
\examples{
\dontrun{lambda1 <- 0.01
lambda2 <- 500*exp(-(1:150)/5)
result <- fusedepoc(Y=list(Y1, Y2), U=list(U1, U2), 
 lambda1=lambda1,lambda2=lambda2, method='A')
}

}
