% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2json.R
\name{fusedepocToJSON}
\alias{fusedepocToJSON}
\title{Creates a JSON string from fused EPoC results usable in the
GUI}
\usage{
fusedepocToJSON(result, species, netnames)
}
\arguments{
\item{result}{Fused EPoC result}

\item{species}{vector of names of species / genes}

\item{netnames}{vector of names of cancers}
}
\value{
a JSON string comprising of all that is needed for
the GUI
}
\description{
Creates a JSON string from fused EPoC results usable in the
GUI
}
\examples{

D <- syntheticPerturbZap(1000,10)
lambda1 <- seq(from=0.01, to=1, length.out=10)
lambda2 <- 500*exp(-(1:150)/5)
result <- fusedepoc(Y=list(D$r1$Y, D$r2$Y), U=list(D$r2$U,
  D$r2$U), lambda1=lambda1, lambda2=lambda2, method='A')

result.json <- fusedepocToJSON(result, paste(1:10),
  c('type A','type B'))

}
