% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusingdistrib.R
\name{fuseplot}
\alias{fuseplot}
\title{Show distribution of fusing over genes}
\usage{
fuseplot(result, ks, l1 = 1)
}
\arguments{
\item{result}{Fused EPoC result set}

\item{ks}{the two classes to compare}

\item{l1}{index of lambda1 vector}
}
\description{
Plot a heatmap showing how fusing is distributed over genes
and lambda2. This can be used for a broad overview to identify genes 
behaving similar or different in two classes.
}
\examples{
D <- syntheticPerturbZap(1000,10)
lambda1 <- 0.01
lambda2 <- 500*exp(-(1:150)/5)
result <- fusedepoc(Y=list(D$r1$Y, D$r2$Y), U=list(D$r2$U,D$r2$U), 
  lambda1=lambda1,lambda2=lambda2, method='A')
H <- fuseplot(result, c(1,2), l1=1)
}
