% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoc_fused_graph.R
\name{getGraph}
\alias{getGraph}
\title{Get colored multi-graph from Fused EPoC results}
\usage{
getGraph(i, j, A, labs = NULL, threshIsLink = 0.01,
  threshIsSame = 0.001)
}
\arguments{
\item{i}{lambda1 index (LASSO penalty)}

\item{j}{lambda2 index (Fused penalty)}

\item{A}{Fused EPoC result set or a coefficient array}

\item{labs}{The names of genes/species/nodes}

\item{threshIsLink}{threshold for abs(coefficient) to be
included as a link}

\item{threshIsSame}{threshold for difference between
coefficients to be fused}
}
\description{
Given parameter indices for lambda1 and lambda2, get a
multi-graph from a result set of Fused EPoC
}
