% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoc_fused.R
\name{netsize}
\alias{netsize}
\title{Net size}
\usage{
netsize(A, thresh = 0.01)
}
\arguments{
\item{A}{a network matrix}

\item{thresh}{coefficients with a magnitude smaller than
threshold will not be counted}
}
\description{
This function show the number of links for an individual network matrix
}
\keyword{links}
\keyword{network}
