% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused.R
\name{penmatrix}
\alias{penmatrix}
\title{Penalty matrix}
\usage{
penmatrix(p, K)
}
\arguments{
\item{p}{The number of predictors}

\item{K}{The number of classes}
}
\description{
This function allows you to create a penalty matrix
}
\keyword{penalty}
