% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2json.R
\name{savejson}
\alias{savejson}
\title{Save a JSON string as a file}
\usage{
savejson(json, filename = NULL, format = "jsmodule")
}
\arguments{
\item{json}{JSON string}

\item{filename}{filename}

\item{format}{jsmodule, json or js}
}
\description{
Save a JSON string as a file
}
