% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticData.R
\name{syntheticPerturbZap}
\alias{syntheticPerturbZap}
\title{Generate synthetic data for Fused EPoC}
\usage{
syntheticPerturbZap(N, p)
}
\arguments{
\item{N}{number of samples}

\item{p}{number of nodes = genes in the Gene Regulatory
Network}
}
\description{
This function allows you to create randomized synthetic
data which can be used by Fused EPoC. Two networks are
generated where the second is a modified version of the
first with both perturbed and zapped links.
}
\examples{
D <- syntheticPerturbZap(1000,10)
lambda1 <- seq(from=0.01, to=1, length.out=10)
lambda2 <- 500*exp(-(1:150)/5)
result <- fusedepoc(Y=list(D$r1$Y, D$r2$Y), U=list(D$r2$U,
D$r2$U), lambda1=lambda1, lambda2=lambda2, method='A')
gl <- getGraph(2,40,result)
\dontrun{Rgraphviz::plot(gl$ng)}
\dontrun{graph::plot(gl$graph)}
}
